+++ src/test/groovy/ru/vyarus/guice/persist/orient/base/support/EmptyAutoScanModule.groovy
@@ -1,8 +1,9 @@
 package ru.vyarus.guice.persist.orient.base.support
 
 import com.google.inject.AbstractModule
+import ru.vyarus.guice.persist.orient.OrientModule
+import ru.vyarus.guice.persist.orient.support.AutoScanSchemeModule
 import ru.vyarus.guice.persist.orient.support.Config
-import ru.vyarus.guice.persist.orient.support.AutoScanSchemeOrientModule
 
 /**
  * Module with schema init from classpath scanned objects but with wrong package configured (no objects will be found)
@@ -13,6 +14,7 @@ class EmptyAutoScanModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        install(new AutoScanSchemeOrientModule(Config.DB, Config.USER, Config.PASS, "wrong.package"))
+        install(new OrientModule(Config.DB, Config.USER, Config.PASS))
+        install(new AutoScanSchemeModule("wrong.package"))
     }
 }