+++ src/test/groovy/ru/vyarus/guice/persist/orient/base/support/EmptyPackageModule.groovy
@@ -1,8 +1,9 @@
 package ru.vyarus.guice.persist.orient.base.support
 
 import com.google.inject.AbstractModule
+import ru.vyarus.guice.persist.orient.OrientModule
 import ru.vyarus.guice.persist.orient.support.Config
-import ru.vyarus.guice.persist.orient.support.PackageSchemeOrientModule
+import ru.vyarus.guice.persist.orient.support.PackageSchemeModule
 
 /**
  * Module with schema init from package but wrong package specified (no entities will be mapped)
@@ -13,6 +14,7 @@ class EmptyPackageModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        install(new PackageSchemeOrientModule(Config.DB, Config.USER, Config.PASS, "wrong.package"))
+        install(new OrientModule(Config.DB, Config.USER, Config.PASS))
+        install(new PackageSchemeModule("wrong.package"))
     }
 }