+++ src/test/groovy/ru/vyarus/guice/persist/orient/support/modules/PackageSchemeModule.groovy
@@ -1,8 +1,8 @@
 package ru.vyarus.guice.persist.orient.support.modules
 
 import com.google.inject.AbstractModule
+import ru.vyarus.guice.persist.orient.OrientModule
 import ru.vyarus.guice.persist.orient.support.Config
-import ru.vyarus.guice.persist.orient.support.PackageSchemeOrientModule
 
 /**
  * Module with predefined scheme mapping from objects in package.
@@ -13,6 +13,7 @@ class PackageSchemeModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        install(new PackageSchemeOrientModule(Config.DB, Config.USER, Config.PASS, Config.MODEL_PKG))
+        install(new OrientModule(Config.DB, Config.USER, Config.PASS))
+        install(new ru.vyarus.guice.persist.orient.support.PackageSchemeModule(Config.MODEL_PKG))
     }
 }